+++ packages/runtime/src/storage.ts
@@ -6,9 +6,9 @@ import {
 	writeFileSync,
 } from './fs';
 import { URL } from './polyfills/url';
+import { Profile } from './switch/profile';
 import { Application } from './switch/ns';
 import { INTERNAL_SYMBOL } from './internal';
-import { currentProfile } from './switch/profile';
 import {
 	assertInternalConstructor,
 	createInternal,
@@ -16,7 +16,6 @@ import {
 	def,
 	encodeUTF16,
 } from './utils';
-import type { FsDev } from './switch/fsdev';
 
 interface StorageImpl {
 	clear(): void;
@@ -125,26 +124,27 @@ Object.defineProperty(globalThis, 'localStorage', {
 	configurable: true,
 	get() {
 		const { self } = Application;
-		const profile = currentProfile({ required: true });
-
-		let dev: FsDev;
-		try {
-			dev = self.mountSaveData(profile);
-		} catch (err: any) {
-			// rethrow if not `ResultTargetNotFound` (meaning save data has not been created)
-			if (err.description !== 1002) throw err;
-
-			self.createSaveData(profile);
-			dev = self.mountSaveData(profile);
+		let profile = Profile.current;
+		while (!profile) {
+			profile = Profile.current = Profile.select();
 		}
 
-		const base = new URL('localStorage/', dev.url);
+		let saveData = self.findSaveData({
+			type: 1 /* FsSaveDataType_Account */,
+			uid: profile.uid,
+		});
+		if (!saveData) {
+			saveData = self.createProfileSaveDataSync(profile);
+		}
+		const base = new URL('localStorage/', saveData.mount());
+
 		const keyToPath = (key: any) => {
 			const url = `${base.href}_${Array.from(String(key))
 				.map((l) => l.charCodeAt(0).toString(16))
 				.join('_')}`;
 			return url;
 		};
+
 		const pathToKey = (path: string) => {
 			const key = String.fromCharCode(
 				...path
@@ -158,7 +158,7 @@ Object.defineProperty(globalThis, 'localStorage', {
 		const impl: StorageImpl = {
 			clear() {
 				removeSync(base);
-				dev.commit();
+				saveData!.commit();
 			},
 			getItem(key: string): string | null {
 				const b = readFileSync(keyToPath(key));
@@ -174,11 +174,11 @@ Object.defineProperty(globalThis, 'localStorage', {
 			},
 			removeItem(key: string): void {
 				removeSync(keyToPath(key));
-				dev.commit();
+				saveData!.commit();
 			},
 			setItem(key: string, value: string): void {
 				writeFileSync(keyToPath(key), encodeUTF16(String(value)));
-				dev.commit();
+				saveData!.commit();
 			},
 			length(): number {
 				const keys = readDirSync(base);