+++ packages/runtime/src/switch/ns.ts
@@ -1,8 +1,12 @@
 import { $ } from '../$';
-import { FsDev } from './fsdev';
-import { crypto } from '../crypto';
+import {
+	SaveData,
+	SaveDataCreationInfoWithNacp,
+	SaveDataFilter,
+} from './savedata';
+import { inspect } from '../inspect';
 import { readFileSync } from '../fs';
-import { proto, stub } from '../utils';
+import { first, proto, stub } from '../utils';
 import type { Profile } from './profile';
 
 let init = false;
@@ -14,7 +18,7 @@ function _init() {
 	init = true;
 }
 
-const genName = () => `s${crypto.randomUUID().replace(/-/g, '').slice(0, 16)}`;
+const getSaveDataOwnerId = (nacp: DataView) => nacp.getBigUint64(0x3078, true);
 
 /**
  * Represents an installed application (game) on the console,
@@ -83,7 +87,7 @@ export class Application {
 	 *
 	 * ```typescript
 	 * const nro = await Switch.readFile('sdmc:/hbmenu.nro');
-	 * const app = Switch.Application.fromNro(nro);
+	 * const app = new Switch.Application(nro);
 	 * console.log(app.name);
 	 * ```
 	 *
@@ -104,6 +108,10 @@ export class Application {
 		stub();
 	}
 
+	createSaveDataSync(info: SaveDataCreationInfoWithNacp) {
+		return SaveData.createSync(info, this.nacp);
+	}
+
 	/**
 	 * Creates the Save Data for this {@link Application} for the provided user profile.
 	 *
@@ -116,12 +124,16 @@ export class Application {
 	 *
 	 * @param profile The {@link Profile} to create the save data for.
 	 */
-	createSaveData(profile: Profile) {
-		$.fsdevCreateSaveData(
-			1, // FsSaveDataType_Account
-			this.nacp,
-			profile.uid,
-		);
+	createProfileSaveDataSync(
+		profile: Profile,
+		info?: Partial<SaveDataCreationInfoWithNacp>,
+	) {
+		return this.createSaveDataSync({
+			spaceId: 1 /* FsSaveDataSpaceId_User */,
+			type: 1 /* FsSaveDataType_Account */,
+			...info,
+			uid: profile.uid,
+		});
 	}
 
 	/**
@@ -129,54 +141,32 @@ export class Application {
 	 *
 	 * @param index The save index ID. Defaults to `0`.
 	 */
-	createCacheData(index = 0) {
-		$.fsdevCreateSaveData(
-			5, // FsSaveDataType_Cache
-			this.nacp,
-			[0n, 0n],
+	createCacheSaveDataSync(
+		index = 0,
+		info?: Partial<SaveDataCreationInfoWithNacp>,
+	) {
+		return this.createSaveDataSync({
+			spaceId: 1 /* FsSaveDataSpaceId_User */,
+			type: 5 /* FsSaveDataType_Cache */,
+			...info,
 			index,
-		);
+		});
+	}
+
+	*filterSaveData(filter?: Omit<SaveDataFilter, 'applicationId'>) {
+		const nacp = new DataView(this.nacp);
+		yield* SaveData.filter({
+			applicationId: getSaveDataOwnerId(nacp),
+			...filter,
+		});
+	}
+
+	findSaveData(filter: Omit<SaveDataFilter, 'applicationId'>) {
+		return first(this.filterSaveData(filter));
 	}
 
 	/**
-	 * Mounts the save data for this application such that filesystem operations may be used.
-	 *
-	 * @example
-	 *
-	 * ```typescript
-	 * const profile = Switch.currentProfile({ required: true });
-	 * const saveData = app.mountSaveData(profile);
-	 *
-	 * // Use the filesystem functions to do operations on the save mount
-	 * console.log(Switch.readDirSync(saveData.url));
-	 *
-	 * // Make sure to use `saveData.commit()` after any write operations
-	 * const saveStateUrl = new URL('state', saveData.url)
-	 * Switch.writeFileSync(saveStateUrl, 'your app state…');
-	 * saveData.commit();
-	 * ```
-	 *
-	 * @param profile The {@link Profile} which the save data belongs to.
-	 * @param name The name of the device mount for filesystem paths. If not provided, a random name is generated.
-	 */
-	mountSaveData(profile: Profile, name = genName()): FsDev {
-		$.fsdevMountSaveData(name, this.nacp, profile.uid);
-		return new FsDev(name);
-	}
-
-	/**
-	 * Mounts the Cache storage for this application such that filesystem operations may be used.
-	 *
-	 * @param index The save index ID. Defaults to `0`.
-	 * @param name The name of the device mount for filesystem paths. If not provided, a random name is generated.
-	 */
-	mountCacheData(index = 0, name = genName()): FsDev {
-		$.fsdevMountCacheStorage(name, this.nacp, index);
-		return new FsDev(name);
-	}
-
-	/**
-	 * An `Application` instance representing the currently running application.
+	 * An {@link Application} instance representing the currently running application.
 	 */
 	static get self(): Application {
 		if (!self) {
@@ -201,3 +191,8 @@ export class Application {
 	}
 }
 $.nsAppInit(Application);
+
+Object.defineProperty(Application.prototype, inspect.keys, {
+	enumerable: false,
+	value: () => ['id', 'nacp', 'icon', 'name', 'version', 'author'],
+});