+++ packages/create-nxjs-app/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-nxjs-app
 
+## 0.0.42
+
 ## 0.0.41
 
 ### Patch Changes