+++ packages/patch-nacp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nx.js/patch-nacp
 
+## 0.0.2
+
+### Patch Changes
+
+- Support `productName` in "package.json" for NACP title value ([`54d990bb3217ae325ff740851325ea62a607d271`](https://github.com/TooTallNate/nx.js/commit/54d990bb3217ae325ff740851325ea62a607d271))
+
 ## 0.0.1
 
 ### Patch Changes